Created ViewModel-s
Created CardViewModel (controller that contains functions for working with cards) and ProjectViewController (controller that contains functions for working with an array of projects and with projects in general).
Added the ability to create new projects (select fields that will be in cards).
Moved all functions that were in models to ViewModel and some functions from View.